Convert ColumnDropdown and ItemsPerPageDropdown to typescript
